+++ glide-core/src/socket_listener.rs
@@ -365,6 +365,7 @@ fn get_command(request: &Command) -> Option<Cmd> {
         RequestType::Rename => Some(cmd("RENAME")),
         RequestType::DBSize => Some(cmd("DBSIZE")),
         RequestType::Brpop => Some(cmd("BRPOP")),
+        RequestType::PfAdd => Some(cmd("PFADD")),
     }
 }