+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -18,6 +18,9 @@ public class TransactionTestUtilities {
     private static final String key6 = "{key}" + UUID.randomUUID();
     private static final String key7 = "{key}" + UUID.randomUUID();
     private static final String key8 = "{key}" + UUID.randomUUID();
+    // TODO rename after #160
+    private static final String hllKey1 = "{key}:hllKey1-" + UUID.randomUUID();
+    private static final String hllKey2 = "{key}:hllKey2-" + UUID.randomUUID();
     private static final String value1 = UUID.randomUUID().toString();
     private static final String value2 = UUID.randomUUID().toString();
     private static final String value3 = UUID.randomUUID().toString();
@@ -91,6 +94,8 @@ public class TransactionTestUtilities {
 
         baseTransaction.configResetStat();
 
+        baseTransaction.pfadd(hllKey1, new String[] {"a", "b", "c"});
+
         return baseTransaction;
     }
 
@@ -141,7 +146,8 @@ public class TransactionTestUtilities {
             2L,
             OK,
             Map.of("timeout", "1000"),
-            OK
+            OK,
+            1L, // pfadd(hllKey1, new String[] {"a", "b", "c"})
         };
     }
 }